requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.6.1
Django==2.1.7
django-appconf==1.0.3
django-autofixture==0.12.1
django-concurrency==1.4
django-cryptography==0.3
django-encrypted-model-fields==0.5.8
django-extensions==2.1.6
django-filter==2.1.0
django-guardian==1.5.0
django-js-asset==1.2.2
django-mptt==0.9.1
django-polymorphic==2.0.3
django-redis==4.10.0
django-rest-framework==0.1.0
django-rest-polymorphic==0.1.8
djangorestframework==3.9.2
djangorestframework-simplejwt==4.1.0
idna==2.8
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
pyasn1==0.4.5
pycparser==2.19
pycrypto==2.6.1
PyJWT==1.7.1
PyMySQL==0.9.3
python3-keyczar==0.71rc0
pytz==2018.9
redis==3.2.0
requests==2.21.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.24.1
